+++ dat/missions/baron/baron_baron.lua
@@ -4,9 +4,20 @@
  <unique />
  <priority>4</priority>
  <chance>100</chance>
- <location>None</location>
+ <location>bar</location>
+ <faction>Empire</faction>
+ <faction>Dvaered</faction>
+ <faction>Sirius</faction>
+ <cond>
+   if var.peek("baron_hated") then
+      return false
+   end
+   if player.wealth() &lt; 1e6 then
+      return false
+   end
+   return require("misn_test").reweight_active()
+ </cond>
  <notes>
-  <done_evt name="Baroncomm_baron">Triggers</done_evt>
   <campaign>Baron Sauterfeldt</campaign>
  </notes>
 </mission>
@@ -19,6 +30,7 @@ local fmt = require "format"
 local baron = require "common.baron"
 local vni = require "vnimage"
 local ccomm = require "common.comm"
+local neu = require "common.neutral"
 
 local pnt, sys1 = spob.getS("Varia")
 local sys2 = system.get("Ingot")
@@ -30,6 +42,8 @@ local _agent1img, agent1prt = vni.pirate()
 local _agent2img, agent2prt = vni.pirate()
 local _agent3img, agent3prt = vni.pirate()
 
+local npc_name = _("Unaffiliated Pilot")
+local npc_image, npc_portrait = vni.pirate()
 function create ()
    -- Can do inclusives as pilot.clear() is not called
    if not misn.claim({ sys1, sys2 }, true) then
@@ -37,7 +51,56 @@ function create ()
       abort()
    end
 
-   -- Mission should be accepted already from the event
+   misn.setNPC( npc_name, npc_portrait, _("A pilot that seems to be looking for someone to do something for them.") )
+end
+
+function accept ()
+   local accepted = false
+
+   vn.clear()
+   vn.scene()
+   local plt = vn.newCharacter( npc_name, {image=npc_image} )
+   vn.transition()
+
+   vn.na(_([[You're greeted by the nondescript pilot who doesn't seem to be affiliated with anyone you know.]]))
+   plt(_([["Hello there! I represent a man by the name of Baron Sauterfeldt. You may have heard of him in your travels? No? Well, I suppose you can't have it all. My employer is a moderately influential man, you see, and... But no, I'll not bore you with the details. The bottom line is, Lord Sauterfeldt is looking for hired help, and you seem like the sort he needs, judging by your ship."]]))
+   plt(_([[You inquire what it is exactly this Mr. Sauterfeldt needs from you. "Oh, nothing too terribly invasive, I assure you. His Lordship currently needs a courier, nothing more. Erm, well, a courier who can't be traced back to him, if you understand what I mean. So what do you think? Sound like a suitable job for you? The pay is good, I can assure you that!"]]))
+   vn.na(_([[You pause for a moment before responding to this sudden offer. It's not everyday that people come to you with work instead of you looking for it, but then again this job sounds like it could get you in trouble with the authorities. What will you do?]]))
+   vn.menu{
+      {_("Accept the job"), "accept"},
+      {_("Politely decline"), "decline"},
+      {_("Angrily refuse"), "refuse"},
+   }
+
+   vn.label("decline")
+   plt(_([["Oh. Oh well, too bad. I'll just try to find someone who will take the job, then. Sorry for taking up your time. See you around!"]]))
+   vn.done()
+
+   vn.label("refuse")
+   plt(_([[The pilot frowns. "I see I misjudged you. I thought for sure you would be more open-minded. Get out of my sight and never show your face to me again! You are clearly useless to my employer."]]))
+   vn.func( function ()
+      var.push("baron_hated", true)
+      neu.addMiscLog( _([[You were offered a sketchy-looking job by a nondescript pilot, but you angrily refused to accept the job. It seems whoever the pilot worked for won't be contacting you again.]]) )
+   end )
+   vn.done()
+
+   vn.label("accept")
+   vn.func( function () accepted=true end )
+   plt(fmt.f(_([["Oh, that's great! Okay, here's what Baron Sauterfeldt needs you to do. You should fly to the Dvaered world {pnt}. There's an art museum dedicated to one of the greatest Warlords in recent Dvaered history. I forget his name. Drovan or something? Durvan? Uh, anyway. This museum has a holopainting of the Warlord and his military entourage. His Lordship really wants this piece of art, but the museum has refused to sell it to him. So, we've sent agents to... appropriate... the holopainting."]]),
+      {pnt=pnt}))
+   vn.na(_([[You raise an eyebrow, but the pilot on the other end seems to be oblivious to the gesture.]]))
+   plt(fmt.f(_([["So, right, you're going to {pnt} to meet with our agents. You should find them in the spaceport bar. They'll get the item onto your ship, and you'll transport it out of Dvaered space. All quiet-like of course. No need for the authorities to know until you're long gone. Don't worry, our people are pros. It'll go off without a hitch, trust me."]]),
+      {pnt=pnt}))
+   vn.na(_([[You smirk at that. You know from experience that things seldom 'go off without a hitch', and this particular plan doesn't seem to be all that well thought out. Still, it doesn't seem like you'll be in a lot of danger. If things go south, they'll go south well before you are even in the picture. And even if the authorities somehow get on your case, you'll only have to deal with the planetary police, not the entirety of House Dvaered.]]))
+   vn.na(_([[You ask the Baron's messenger where this holopainting needs to be delivered.]]))
+   plt(fmt.f(_([["His Lordship will be taking your delivery in the {sys} system, aboard his ship, the Pinnacle," he replies. "Once you arrive with the holopainting onboard your ship, hail the Pinnacle and ask for docking permission. They'll know who you are, so you should be allowed to dock. You'll be paid on delivery. Any questions?"]]),
+      {sys=sys2}))
+   vn.na(fmt.f(_([[You indicate that you know what to do, then cut the connection. Next stop: planet {pnt}.]]),
+      {pnt=pnt}))
+   vn.run()
+
+   if not accepted then return end
+
    misn.accept()
 
    misn.setTitle(_("Baron"))